+++ src/users/users.controller.ts
@@ -16,7 +16,7 @@ import { ChangePasswordDto } from './dto/password.dto';
 import { UpdateResult } from 'typeorm';
 import { SaveArticleDto } from '../articles/dto/articles.dto';
 import { Article } from '../articles/entities/articles.entity';
-import { UsersToArticles } from '../userstoarticles/userstoarticles.entity';
+import { UsersToArticles } from '../userstoarticles/entities/userstoarticles.entity';
 import { UserSerialize } from './dto/users.serializer';
 import { AuthGuard } from '@nestjs/passport';
 import { ShareArticleDto } from './dto/shareArticle.dto';
