+++ src/userstoarticles/userstoarticles.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersToArticles } from './userstoarticles.entity';
+import { UsersToArticles } from './entities/userstoarticles.entity';
 import { UsersToArticlesService } from './userstoarticles.service';
 
 @Module({
